+++ src/traits/mod.rs
@@ -25,9 +25,11 @@
 pub mod error_category;
 pub mod error_ops;
 pub mod into_error_context;
+pub mod result_ext;
 pub mod with_error;
 
 pub use error_category::ErrorCategory;
 pub use error_ops::ErrorOps;
 pub use into_error_context::IntoErrorContext;
+pub use result_ext::{BoxedResultExt, ResultExt};
 pub use with_error::WithError;